scx_flash: Always prefer reusing the same CPU
When the primary scheduling domain is full, tasks may overflow to CPUs
outside of it. On wakeup, we aggressively attempt to pull these tasks
back into the primary domain. However, if the primary domain is
saturated, this can lead to excessive task migrations.

To mitigate this, always allow reusing the same CPU during task wakeup.

This change results in a +20–25% performance improvement in MySQL
read/write benchmarks.

Signed-off-by: Andrea Righi <arighi@nvidia.com>