scx: add basic map implementation for unit testing
In order to unit test most bpf things we will need to have a map
implementation that we can work with.  This is a bare bones map
implementation, it's not meant to be fast, just correct.

We then #define the equivalent bpf function names so they can be
replaced with our own implementation.

In order to avoid re-writing a bunch of BPF code to be testable, we keep
track of the pointers to the map structs and register those with our
scx_test_map versions in the test code, so all of the bpf map helpers
can be used unchanged and then we just look up the pointer to find the
corresponding scx_test_map.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>