scx: rework p2dq bpf unittest and add a new one
Instead of adding the unittest code directly to the bpf file, we can
just include the bpf file into a special test file that we can use to do
all the special test setup stuff.

With this we can add the supporting code to take advantage of the
map support and add a test for lookup_cpu_ctx, which is used in a
variety of places.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>