scx_flash: Introduce --no-builtin-idle
Rely on the in-kernel built-in idle CPU selection policy by default, if
available and introduce an option to explicitly disable it and fallbak
to the custom idle CPU selection policy.

Since the scheduler relies on the new kfunc scx_bpf_select_cpu_and(),
this option has no effect with kernels that don't provide this API
(kernels < 6.16).

Signed-off-by: Andrea Righi <arighi@nvidia.com>