scx_lavd: Be conservative about migrating to a waker's CPU.
Previously, when selecting an idle CPU in the event of a synchronous wake-up,
we treated the previous CPU and the waker's CPU equally, so triggering more
task migrations.

Now, we treat them differently; we prefer to reuse the previous CPU and its
domain. If that's not available, we will consider using waker's domain (not
the CPU).

Additionally, when there is no fully idle CPU in the system, we won't chase
a partially idle CPU in neighboring domains.

Signed-off-by: Changwoo Min <changwoo@igalia.com>