scx_flash: Use per-CPU DSQs in function of CPU utilization
Rather than limiting CPU affinity to just interactive tasks, re-enqueue
tasks to the per-CPU DSQ as long as the CPU is not saturated.

Define saturation as exceeding 50% CPU usage: above this threshold,
tasks are redirected to the per-node DSQ to enable migration and improve
work conservation and overall system responsiveness.

Tested-by: Piotr Gorski <piotrgorski@cachyos.org>
Signed-off-by: Andrea Righi <arighi@nvidia.com>