scs_lavd: Partially inherit the waker's latency criticality.
We had inherited waker's latency criticality when the waker's latency is
higher than the wakee's. This is good in the sense that the wakee has
some context for who woke me up. However, when two tasks are tightly
interrelated, the latency criticality inheritance mechanism makes the
criticality of all involved tasks the same. Hence, change to add some
additional credit from the waker, so the latency criticality values are
still different among tightly interrelated tasks.

Signed-off-by: Changwoo Min <changwoo@igalia.com>