vmlinux: generate using a predictable config
Currently vmlinux.h files are generated with a config chosen by the
maintainer. Other developers wishing to do this need to create their
own, which is non-trivial, because some non-obvious config options are
mandatory for sched_ext.

Furthermore, after generating vmlinux.h for one architecture, the
.config file is reused for the next one. Reusing .config like this
leads to loss of config options. For example, s390 ends up being
compiled without NUMA.

Unless .config exists, which is an indication to keep the current
behavior, use the architecture's defconfig with a few extra options on
top.

Signed-off-by: Ilya Leoshkevich <iii@linux.ibm.com>