chaos: re-enable on sched_ext/for-next kernel
Chaos was disabled on the sched_ext/for-next kernel a while ago because
of RCU stalls that appeared after a kernel update. There have been many changes
and fixes to chaos, particularly around the p2dq interrop recently, so
let's try running on this kernel again as it's the most useful to track.

Closes #2272

Test plan:
- CI, several times.