scx_lavd: Boost the time slice of long-running tasks whenever possible.
If the task's avg_runtime is greater than the regular time slice
(i.e., taskc->avg_runtime > sys_stat.slice), that means the task could be
scheduled out due to a shorter time slice than required. In this case,
let's consider boosting task's time slice. However, if there are pinned
tasks waiting to run on this CPU, we do not boost the task's time slice
to avoid delaying the pinned task that cannot be run on another CPU.

When the system is not heavily loaded, so it can serve all tasks within
the targeted latency (slice_max_ns <= sys_stat.slice), we fully boost
task's time slice. Let's set the task's time slice to its avg_runtime
(+ some bonus) to reduce unnecessary involuntary context switching. Even
in this case, we want to limit the maximum time slice to
LAVD_SLICE_BOOST_MAX (not infinite) because we want to revisit if the
task is placed on the best CPU at least every LAVD_SLICE_BOOST_MAX interval.

When the system is under high load, we will boost the time slice of only
latency-critical tasks, which are likely in the middle of a task chain.
Also, increase the time slice proportionally to the latency criticality up
to 2x the regular time slice.

If slice boost is either not possible, not necessary, or not eligible,
assign the regular time slice.

Signed-off-by: Changwoo Min <changwoo@igalia.com>