scx_lavd: Keep running task's flag on cpu_ctx.
It is necessary to know the running task's flag on a CPU. So, let's keep
the copy of the task's flag on the cpu_ctx structure. As the flag is kept
on the cpu_ctx, it is not necessary to separately manage a boolean flag,
lock_holder, so remove it.

There is no functional change.

Signed-off-by: Changwoo Min <changwoo@igalia.com>