scx_lavd: Preempt a slice-boosted task if appropriate.
When a slice-boosted task is running on a chosen CPU by the ops.enqueue()
path, check if it is appropriate to preempt the slice-boosted task.

If the enqueuing task is more latency-critical, let's cancel the slice
boosting and kick the boosted task immediately.

If the enqueuing task is affinitized, it should run on the preferred CPU.
To avoid the affinitized task being delayed too long, let's cancel the slice
boosting so that the currently running task on the preferred CPU is scheduled
out, just like other regular tasks.

In addition to the preemption path at ops.enqueue(), check if there are
pinned tasks that are waiting to run on a particular CPU at ops.tick() and
the sys_stat interval. If there is, shrink the boosted time slice into the
regular one to avoid pinned tasks being delayed indefinitely.

Signed-off-by: Changwoo Min <changwoo@igalia.com>