scx_lavd: More accurate accounting of CPU utilization.
Previously, CPU utilization was updated only when a task transitioned from
a running to a stopped state. That was OKish when the maximum time slice
(e.g., 5 msec) is smaller than the sys_state interval (e.g., 20 msec).
However, a task execution can span across sys_stat interval boundaries.
Moreover, in the case of a long-running task that runs longer than a
sys_state interval, the CPU utilization would be measured at 0% even
if its actual utilization is 100%.

To avoid such a problem and to accurately account for CPU utilization,
especially for long-running tasks, we now accumulate CPU and utilization
of running tasks at ops.stopping(), every ops.tick(), and ops.dispatch()
when prev_task is rescheduled.

In this way, we can keep all the update operations in a local CPU and bound
the maximum error by the tick interval.

Signed-off-by: Changwoo Min <changwoo@igalia.com>