nix: switch to fenix to better pin rust dependencies
Currently CI and anyone using the Nix flake gets their rust dependencies
from nixpkgs. This is generally fine as they are locked, but we have
less control over which version is selected than with tools like rustup.

Switch to fenix[1] which gives equivalent control to rustup. This
enables testing on nightly toolchains with a simple change. It's mostly
an ergonomic benefit for now. Required some minor clippy fixes for the
updated clippy version.

Test plan:
- CI

[1] https://github.com/nix-community/fenix