libbpf-cargo: place BPF objects in its own section and look up for veristat
Currently the approach for veristat is to look at the build directory of
a scheduler and extract any potentially matching objects. This is not
ideal as we can generate excessive matches or miss files if the build
process changes.

libbpf/libbpf-rs#1197 modifies libbpf-cargo to place the body of the
.bpf.o file in a specifically named section ".bpf.objs". This ends up in
the final binary as long as it is used and can be readily extracted (the
content was already in the binary but hard to locate).

Use this new feature and the script added in
`scripts/extract_bpf_objects.sh` to replace the `find` command and
dependency on the build directory in the meson veristat job & hence the
CI. Move the CI veristat job to be always required as it must succeed.

Also made a similar change to the C scheduler generation, but rather
than upstreaming it this is achieved with `sed` in the Meson build.
Happy to remove this if anyone isn't a fan of the additional build
machinery, but the symmetry is nice. I plan to upstream this to bpftool
eventually but given we support older versions of bpftool and can't pin
the dependency like with Cargo it'll be a while before we see all the
benefits. We don't currently test any C schedulers in the CI so this
isn't overly important.

Test plan:
- CI