ci: use fast-apt-mirror action
Recently the Azure apt mirrors have been horrendous for reliability.
We've added a few new runs for schedulers and are up to 14 integration
tests now, and often we're seeing 2 failing due to apt not downloading
in the 30 minute time.

Use this `fast-apt-mirror` action to select a competent mirror before
starting.

Test plan:
- Ran twice, no failures. The time still isn't completely consistent but
  it seems like a marked improvement.