scx_cosmos: Allow disabling deferred wakeups
The mechanism of doing deferred wakeups using a BPF timer can help
making the hot path in ops.enqueue() faster, but it can also increase
power consumption due to the timer that is contantly checking if there
are tasks waiting on the local DSQs.

This might be problematic on battery-powered devices, so introduce the
option --no-deferred-wakeups to disable this behavior and wake up CPUs
synchronously in ops.enqueue().

This can reduce performance and throughput in certain workloads, but it
helps reducing power consumption.

Example on an AMD Ryzen AI 9 HX 370 laptop:

 - scx_cosmos:
   $ sudo turbostat -S -s PkgWatt,CorWatt
   CorWatt	PkgWatt
   1.99		5.78
   2.00		5.75
   2.01		5.32
   1.99		5.78
   2.01		5.79

 - scx_cosmos --no-deferred-wakeups:
   $ sudo turbostat -S -s PkgWatt,CorWatt
   CorWatt	PkgWatt
   0.11		1.57
   0.14		1.77
   0.17		1.40
   0.14		1.24
   0.17		1.20

Signed-off-by: Andrea Righi <arighi@nvidia.com>