scx_p2dq: Refactor dispatch path for ATQs
When popping from an ATQ in the dispatch path requeue the task to the
local LLC DSQ if the task popped is not the same task as the peeked
task. This fixes issues where ATQs could potentially lose tasks and
cause stalls.

Before doing pick two load balancing always try to consume from the
migration eligible ATQ/DSQ. This gives slightly better throughput at the
LLC level as fewer llc_ctx lookups are needed in the hot path.

Signed-off-by: Daniel Hodges <git@danielhodges.dev>