ci: move veristat job from ubuntu runners to python script
Currently we run veristat when we run a full set of stress tests for a
scheduler. There are a few flaws with this approach:
- Any failures of veristat are completely ignored and still allow the PR
  to merge.
- We spin up an entire virtme-ng VM to run a single veristat invocation,
  even though veristat is very quick and the VMs are quite expensive.
- We only run veristat against the kernel(s) we have chosen for full
  stress tests, which are few because the tests are expensive.

This moves the veristat jobs from the Ubuntu runners to the ci.py Python
script. In doing that, it adds the plumbing for a job internal matrix
and many efficiency improvements. This allows us to run veristat for all
relevant kernels for all schedulers on every PR with a single quick job.

Test plan:
- CI