ci: add test for arighi's patched 6.12
Rustland needs a 6.12 LTS backport to be able to call
`scx_bpf_pick_idle_cpu` from a BPF timer. Add a branch with this
backport to our CI temporarily to confirm it fixes the issue. This will
run 6 hourly and once it's applied to stable we can remove this
additional testing.

Test plan:
- CI for job runs succeeding
- Inspection to confirm the scheduled job runs the same kernel as the
  trailer

CI-Test-Kernel: arighi/6_12