ci: drop arighi-6.12 and apply patch to stable-6.12
We've been A/B testing our tracked stable/6_12 kernel against @arighi's
patched version for a while, and the patch seems to work really well.
Move on from A/B testing to continuous testing where we apply this patch
to the stable/6_12 kernel to check it rebases/remove the red in our CI
until it merges upstream.

Test plan:
- CI, including the 6.12 test that has been failing on rustland without
  the patch

CI-Test-Kernel: stable/6_12