scx_layered: Add layer matching based on hint
Add a new HintEquals matcher attribute for layers that allows changing
layer membership for tasks based on the supplied hint value. Consolidate
all lookup related checks in a lookup_task_hint() helper, and ensure the
value seen lies in the permitted range ([0, 1024]). To ensure the lookup
is a no-op when the --task-hint-map= option is not passed to layered,
make all lookups conditional on a const volatile 'task_hint_map_enabled'
check that resolves to false when the option is not specified.

For now, this will not be synchronized with hint changes in user space,
as they are done using map updates, and potentially in the future
directly over shared memory between the scheduler and the application.
We will address such requirements in subsequent patches.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>