scx_layered: Enable high-frequency layer switching using hints
Introduce support for switching layers dynamically at runtime using
hints at very high frequencies (e.g. at every slice boundary), such
that on each wake up and slice consumption event, the hint is acted
upon and necessary switches are made to keep the task<->hint<->layer
mapping consistent.

This is achieved in a low overhead fashion by first constructing a
lookup table with 1025 entries (for values in the range [0, 1024],
which are valid hints). This is populated after parsing the layer
specifications and fed into Scheduler::init() which then populates
the BPF side lookup map.

On the scheduler side, we refactor switch_to_layer out of
maybe_refresh_layer to contain the necessary bits needed to switch a
task from one layer to another, decrementing and incrementing task
counts of each layer.

We invoke the maybe_refresh_task_layer_from_hint hook from two places in
particular, i.e., wake ups (select_cpu(), enqueue()) and dispatch().

This allows us to process the incoming task on wake up and quickly steer
it into the right layer.  Choosing the layer at wake up is critical as
during low load, picking the idle CPU in the right layer (and thus in
the right cache domain) ends up materially affecting performance
characteristics.

In case of dispatch(), we only test whether a layer membership change is
necessary, and reenqueue() the task, so that it gets processed for a
change. This is because for runtime accounting on stopping(), we'd want
it to be charged to the task's layer prior to the switch.

From kernel 6.11 onwards, stopping() isn't called at slice boundaries.
If dispatch() decides to continue a task repeatedly on a CPU (e.g., due
to availability of idle CPUs elsewhere on the machine), we need to be
able to change layer membership if needed and reenqueue it on the new
layer's CPU if necessary. With SCX_OPS_ENQ_LAST set, we always get
control when a slice expires without the kernel deciding to continue
running previous task on our behalf, thus we have an opportunity to
change the layer membership at all slice boundaries (regardless of if
we reached dispatch() due to slice expiration (SCX_TASK_QUEUED = true),
or the task going off the run queue (SCX_TASK_QUEUED = false)).

Therefore, we only need to handle all the cases in dispatch(), and
reenqueue the task when the right conditions hold.

The vtime adjustment needs some special care as the timelines are not
directly comparable across layer DSQs; this will be addressed in a
separate patch set.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>