[SCX][lavd] modify lavd to use new peek operation
Summary:

The per-cpu queues mode in LAVD checks multiple queues for potential steal targets.
These checks can be made more efficient with the forthcoming O(1) lockless peek operation.
This patch updates LAVD to use peek in three places:

- try_to_steal: picking cpu with lowest vtime
- on consume_task: checking the cpu DSQ
- on consume_task: checking the domain DSQ

Note that these usages are necessarily racy because no lock is held. That is, we
could peek a low vtime, but the task may have moved or completed by the time we
attempt to steal it.

Test Plan:
 - ran schbench under a VM [1 cpdom, 50cpus]

Reviewers: daidavid