scx_bpf_unittests: fix build and run tests again
These tests were disabled recently because of missing arena stubs. Add
them as stubs (for now) and re-enable the tests.

I want to start using these tests in scx_chaos to test for correct
behaviour without the verifier interfering, which will likely require
not stubbing the arena code. But this is good enough for now.

Test plan:
- `cargo test -p scx_bpf_unittests`
- CI