meson: add deprecation warning
Meson as a build system is being replaced by a split between
Cargo/gnumake for Rust/C schedulers. Add a warning before the next
release to ask packagers to switch over to the replacements.

In a future release we'll add an error that can be overridden with a
configuration option, and in the release after that we'll remove the
Meson build entirely.

Test plan:
- CI - the new warning shows up.