scx_lavd: Add virt-llcs support
When enabled, it builds cpdoms based on a virtual set of LLC domains
which is found by searching through a user provided range within a
physical LLC domain while accounting for Big/Little sizes of CPUs.
If no range is provided, it uses a default min/max to search instead.

Signed-off-by: David Dai <david.dai@linux.dev>