scx_lavd: Improve the calculation of performance criticality threshold
Previously, only active CPUs were considered when calculating the performance
criticality threshold, which is used to determine if a task should be assigned
to a big core or a little core. However, in reality, some cores are on anyway
due to pinned tasks. Let's embrace this reality and consider the computing
capacity of overflow CPUs, too.

Signed-off-by: Changwoo Min <changwoo@igalia.com>