nix: move nixpkgs input from JakeHillion to upstream nixpkgs
Currently Nix pulls its nixpkgs input from JakeHillion/nixpkgs/virtme-ng
where I maintain a virtme-ng package. This does work and was done with
the intention of upstreaming this commit, but is a pain to maintain - I
have to push a merge commit to that branch to update nixpkgs in
sched_ext and it's tricky to do atomically.

Grab the virtme-ng package from that branch and vendor it here. This
allows the normal `nix flake update` to work without requiring changes
in a different GitHub repository.

Some refactoring here too, the nix flake is getting a bit too big. Add a
`.nix/pkgs` directory to store the few package definitions. Will move
the flake.nix and flake.lock there as a follow up, it's a bigger change.

Test plan:
- CI