ci: clang-format scx_chaos
In #2718 we added CI checked clang-format for scxtop. The formatting is
okay, and it's improved consistency a lot, especially when it comes to
random contributors or vibe coding.

Move the scxtop .clang-format, which is the kernel's, to the root of the
repository. Mitosis has a custom .clang-format and that will continue
applying.

Test plan:
- CI