scx_layered: Add SystemCpuUtilBelow high-frequency matcher
Implement SystemCpuUtilBelow matcher that checks if system-wide CPU
utilization (10-second EWMA) is below a threshold. This is a matcher
that can be used in conjunction with HintEquals only, and performs the
checks against the threshold in the same wake up paths as the existing
HintEquals check.

A little bit of refactoring is done to introduce a hint_layer_info
struct that carries the hint and associated thresholds, and these are
used to further gate admission if the hint value matches a given layer.

In case a layer is moving in using the hint, and thresholds are
breached, we just simply negate the switching. If the existing task
however matches the layer it is in and exceeds thresholds, we move
it out of the layer. This is only done in the wake up path, and it
is expected that we may not be extremely accurate (at sub-ms scale)
in moving tasks in and out of these layers, but that is acceptable.

The threshold values are compared against EWMA values computed in user
space, to add hysteresis and ensure that oscillation around the
threshold doesn't lead to constant layer bouncing of the task.

Note that the dispatch still does an optimistic check on not continuing
the existing task based on the layer id match, even though the threshold
may not be satisfied. Given that the thresholds are mostly to capture
long term phase changes of the system, this is mostly acceptable, as we
will eventually go through a sleep cycle and detect the change on the
wake up path. This can be further refined in the future if it proves to
be misguided.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>