scx_rustland_core: Fix built-in idle CPU selection on older kernels
Kernels that don't provide scx_bpf_select_cpu_and() allow to use the
built-in idle CPU selection policy only from ops.select_cpu() or
ops.enqueue(). In this case falls back to selecting any idle CPU usable
by the task via scx_bpf_pick_idle_cpu().

This restores compatibility with older kernels while preserving the
optimized built-in idle CPU selection for the user-space scheduler where
supported.

Moreover, when the built-in policy is available, set SCX_WAKE_TTWU,
pretending to be a wakeup, to prioritize faster CPU selection when the
user-space scheduler is actively requesting an idle CPU. Even if in this
case we don't have a real wakeup, the user-space scheduler will likely
release its current CPU, making it a good candidate for dispatching a
task, especially if it's faster than the CPU the task was previously
using.

Fixes: 58464936 ("scx_rustland_core: Rely on built-in idle CPU selection policy")
Signed-off-by: Andrea Righi <arighi@nvidia.com>