scx_layered: Add DsqInsertBelow high-frequency matcher
Similar to the previous commit and following the same ideas, add a
DsqInsertBelow matcher that indexes on the ratio of layer DSQ insertions
instead of the system-wide CPU utilization. Both the previous matcher
and this option can be combined with HintEquals together to provide
multiple levels of protection against layer admission. The overall
system utilization can be low while a layer that is using hints for task
admission is saturated, therefore they can be used for similar purposes
depending on how the system is provisioned, but operate differently.

For latency-critical tasks, it is observed empirically that even a
modest DSQ insertion ratio (e.g., 10% of all wake ups) leads to an
unacceptable spike in perceived latency, therefore if such a situation
is persistent for a while, it would make sense to gate the admission of
tasks to a layer based on this value.

The calculation is done based on existing layer stats, where the ratio
is calculated upon total selections of a task (select_cpu() + enqueue()
idle CPU selection + enqueue() DSQ insertions) in the denominator, and
DSQ insertions in enqueue() as the numerator, and then exponentially
weighted over a 10-second window to smoothen the signal.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>