scx_mitosis: enable clang-format in the CI
Enable checking for clang-format on scx_mitosis in the CI.

Previously mitosis used a different clang-format than the repo so
applying the repo wide one is a bit of churn, but let's do it now for
consistency.

Opted out the struct_ops as the bin packed formatting is disgusting.

Test plan:
- CI