ci: update description when modifying an update-kernels pr
Update kernels commits/PRs include a link to the kernel changelog
automatically. This changes in the commit whenever a PR is updated (when
the PR failed to merge previously), but isn't updated in the PR body.

Add a `gh pr edit` command to ensure the PR body always matches the sole
commit.

Test plan:
- Ran the workflow from this branch
  (https://github.com/sched-ext/scx/actions/runs/18684425487/job/53273179379) on #2889.
  The PR message was updated correctly to match the commit.