scx_mitosis: Make timer_cb more robust
We've seen cases where this rootcg context lookup can fail in a timer.
It feels unlikely for initialization to take so long but in theory it is
possible since the timer is armed in init and there's no synchronization
with the rest. The fix here is to initialize the rootcg storage before
arming the timer so the lookup in the timer can never fail.

Signed-off-by: Dan Schatzberg <schatzberg.dan@gmail.com>