meson: deprecate harder
Now we've had the warning about Meson for a while and everyone we know
about has migrated to pure Cargo, increase the level of the deprecation.
The next step is to hard error the build unless a flag is specified to
override this. This new flag to hide this error is documented and will
force anyone who was taking updates automatically to notice the change.

Test plan:
- CI
- CI with the first version of this commit (no flag). It failed with the
  expected error.