scx_bpfland: Introduce per-CPU DSQs
When targeting specific CPUs with SCX_DSQ_LOCAL_ON in ops.enqueue(),
errors like the following may occur:

  runtime error (SCX_DSQ_LOCAL[_ON] cannot move migration disabled execsnoop[1423] from CPU 9 to 3)

This can happen even when explicitly checking for migration-disabled
tasks or CPU affinity, since a task may be dequeued and re-enqueued
during the enqueue path, changing its affinity or migration-disabled
state in the process.

Because of this, ops.enqueue() is not a reliable place to use
SCX_DSQ_LOCAL_ON and this limitation likely needs to be addressed in the
kernel.

As a temporary workaround, introduce per-CPU DSQs and implement the
proper machinery in ops.dispatch(), where task affinity and migration
state are reliable, to validate dispatches to specific CPUs.

If a task is dispatched to an invalid CPU from ops.enqueue(), it will
automatically be rerouted to the local DSQ of its assigned CPU from
ops.dispatch().

Signed-off-by: Andrea Righi <arighi@nvidia.com>