scx_bpfland: Avoid starvation in per-node DSQs
Tasks queued in per-CPU DSQs are always consumed before those in
per-node DSQs, which can theoretically lead to starvation for tasks
residing in the latter, even if the condition is rare.

To prevent this, compare the deadlines of the first tasks in both queues
and select from the queue with the earliest deadline, ensuring fairer
scheduling between per-CPU and per-node DSQs.

Signed-off-by: Andrea Righi <arighi@nvidia.com>