scx_bpfland: Remove unnecessary DSQ move in ops.dispatch()
Checking task affinity and migration-disabled status in ops.enqueue() is
totally safe.

The problem that commit 10a76808 ("scx_bpfland: Introduce per-CPU DSQs")
claimed to fix was actually due to a missing migration-disabled check in
ops.enqueue().

Now we're doing all the proper checks in ops.enqueue(), so we can drop
the logic that is trying to fix tasks queued to the wrong per-CPU DSQs.

We still need the per-CPU DSQs to prevent the starvation node DSQs
starvation, see commit 8b2bd19e ("scx_bpfland: Avoid starvation in
per-node DSQs"), so we want to keep them for now.

Fixes: 10a76808 ("scx_bpfland: Introduce per-CPU DSQs")
Signed-off-by: Andrea Righi <arighi@nvidia.com>