scx_bpfland: Don't attempt migrations for sticky tasks
When a task is marked as sticky it should always keep running on the
same CPU and it should always override other tasks.

Therefore, dispatch it directly to the same CPU using SCX_DSQ_LOCAL,
instead of attempting to migrate to other CPUs.

This allows handling `stress-ng --sleep 0` stress tests correctly
without stalling when sticky tasks are enabled.

Signed-off-by: Andrea Righi <arighi@nvidia.com>