scx_lavd: gate scx_bpf_dsq_nr_queued call in can_direct_dispatch
When checking whether a CPU should receive a direct dispatch lavd checks
the contents of the per_cpu DSQ to ensure direct dispatching won't
starve that queue. However, this check occurs even when that DSQ does
not exist.

This is confirmed by adding printing of the `scx_bpf_dsq_nr_queued`
result. It's always 4294967294, which is -ENOENT saying the queue does
not exist. `!(-ENOENT)` is false, so `can_direct_dispatch` always
returns false if `per_cpu_dsq` is false.

Modify the check to confirm whether `per_cpu_dsq` is enabled before
checking the size of the queue.

Test plan:
- CI
- `bpf_printk` - before only the `else` branch was hit, now I get a
  mixture.