scx_bpfland: Dynamic time slice
Adjust the time slice in function of the task's priority and the amount
of tasks that are waiting to be dispatched.

This concept was removed a while ago to reduce the scheduling overhead
in presence of a large amount of tasks running, but it can also make the
scheduler less robust, since many tasks piling up in the queues may
increase the list of triggering the stall watchdog and stalls are always
perceived by users as bugs.

Therefore, re-introduce the dynamic time slice to improve robustness.

Signed-off-by: Andrea Righi <arighi@nvidia.com>